a simple HTML game called NinjaMan
similar to PacMan, this game contains a randomly generated maze, a player character, food to collect and gain points, and an enemy to avoid that moves around the screen. My personal complaints with the file as is include: the random maze sometimes creates a maze that prevents the player or enemy from moving around the entire map, and the enemy only moves randomly, rather than "chase" the player.